[slice] move helper class to .cpp file
